Move testPKLookupWithUpdate to TestPerThreadPKLookup.
